set values to 0
